+++ photutils/psf/photometry.py
@@ -9,6 +9,8 @@ from astropy.nddata.utils import overlap_slices
 from astropy.stats import gaussian_sigma_to_fwhm, SigmaClip
 from astropy.table import Table, Column, vstack, hstack
 from astropy.utils.exceptions import AstropyUserWarning
+from astropy.nddata import NDData, StdDevUncertainty
+from astropy.io import fits
 
 from . import DAOGroup
 from .funcs import subtract_psf, _extract_psf_fitting_names
@@ -90,6 +92,9 @@ class BasicPSFPhotometry:
         The radius (in units of pixels) used to compute initial
         estimates for the fluxes of sources. If ``None``, one FWHM will
         be used if it can be determined from the ``psf_model``.
+    noise_calc : callable, optional
+        Function on which uncertainties of data can be calculated. If `None`
+        then the assumption is made that each data point gets unity weighting.
 
     Notes
     -----
@@ -113,7 +118,8 @@ class BasicPSFPhotometry:
     """
 
     def __init__(self, group_maker, bkg_estimator, psf_model, fitshape,
-                 finder=None, fitter=LevMarLSQFitter(), aperture_radius=None):
+                 finder=None, fitter=LevMarLSQFitter(), aperture_radius=None,
+                 noise_calc=None):
         self.group_maker = group_maker
         self.bkg_estimator = bkg_estimator
         self.psf_model = psf_model
@@ -124,6 +130,7 @@ class BasicPSFPhotometry:
         self._pars_to_set = None
         self._pars_to_output = None
         self._residual_image = None
+        self._noise_calc = noise_calc
 
     @property
     def fitshape(self):
@@ -173,10 +180,10 @@ class BasicPSFPhotometry:
 
         Returns
         -------
-        residual_image : 2D array-like, `~astropy.io.fits.ImageHDU`, `~astropy.io.fits.HDUList`
+        residual_image : 2D array-like, `~astropy.nddata.NDData` data attribute
         """
 
-        return self._residual_image
+        return self._residual_image.data
 
     def __call__(self, image, init_guesses=None):
         """
@@ -185,6 +192,22 @@ class BasicPSFPhotometry:
         """
 
         return self.do_photometry(image, init_guesses)
+
+    def _make_image(self, image):
+        """
+        Takes ndarray and returns NDData array with corresponding uncertainty
+        attribute, determined by noise_calc. If noise_calc is `None` then
+        return an empty uncertainty array. Currently assumes for compatibility
+        reasons that these uncertainties are of the standard deviation kind,
+        such that Fitter weights are 1/uncertainty.
+        """
+        if not isinstance(image, NDData):
+            if self._noise_calc is not None:
+                image = NDData(image,
+                               uncertainty=StdDevUncertainty(self._noise_calc(image)))
+            else:
+                image = NDData(image)
+        return image
 
     def do_photometry(self, image, init_guesses=None):
         """
@@ -203,8 +226,11 @@ class BasicPSFPhotometry:
 
         Parameters
         ----------
-        image : 2D array-like, `~astropy.io.fits.ImageHDU`, `~astropy.io.fits.HDUList`
-            Image to perform photometry.
+        image : 2D array-like, `~astropy.io.fits.ImageHDU`, `~astropy.io.fits.HDUList`,
+            or `~astropy.nddata.NDData`
+            Image to perform photometry. If not NDData, the data array will
+            be pulled and passed to noise_data, which will create the
+            corresponding uncertainty array and return an NDData instance.
         init_guesses: `~astropy.table.Table`
             Table which contains the initial guesses (estimates) for the
             set of parameters. Columns 'x_0' and 'y_0' which represent
@@ -229,8 +255,23 @@ class BasicPSFPhotometry:
             uncertanties are not reported.
         """
 
+        # Parse image for the different kinds of accepted inputs; if
+        # not ndarray, convert accordingly.
+        if isinstance(image, fits.HDUList):
+            for i in range(len(image)):
+                if image[i].data is not None:
+                    warnings.warn("Input data is a HDUList object, photometry is "
+                                  "run only for the {0} HDU."
+                                  .format(i), AstropyUserWarning)
+                    image = image[i]
+                    break
+        if isinstance(image, (fits.PrimaryHDU, fits.ImageHDU)):
+            image = image.data
+        image = self._make_image(image)
+
         if self.bkg_estimator is not None:
-            image = image - self.bkg_estimator(image)
+            image = NDData(image.data - self.bkg_estimator(image.data),
+                           uncertainty=StdDevUncertainty(image.uncertainty))
 
         if self.aperture_radius is None:
             if hasattr(self.psf_model, 'fwhm'):
@@ -266,7 +307,7 @@ class BasicPSFPhotometry:
             if self.finder is None:
                 raise ValueError('Finder cannot be None if init_guesses are '
                                  'not given.')
-            sources = self.finder(image)
+            sources = self.finder(image.data)
             if len(sources) > 0:
                 apertures = CircularAperture((sources['xcentroid'],
                                               sources['ycentroid']),
@@ -303,7 +344,7 @@ class BasicPSFPhotometry:
 
         Parameters
         ----------
-        image : numpy.ndarray
+        image : `~numpy.ndarray` or `~astropy.nddata.NDData`
             Background-subtracted image.
         star_groups : `~astropy.table.Table`
             This table must contain the following columns: ``id``,
@@ -318,9 +359,20 @@ class BasicPSFPhotometry:
         -------
         result_tab : `~astropy.table.Table`
             Astropy table that contains photometry results.
-        image : numpy.ndarray
-            Residual image.
+        image : `~astropy.nddata.NDData`
+            Residual image with corresponding uncertainty weights.
         """
+
+        if isinstance(image, NDData):
+            # Pull the values, ignoring the uncertainty type for now
+            try:
+                uncert = image.uncertainty.array
+            except AttributeError:
+                uncert = None
+            image = image.data
+        else:
+            image = image
+            uncert = None
 
         result_tab = Table()
         for param_tab_name in self._pars_to_output.keys():
@@ -346,8 +398,14 @@ class BasicPSFPhotometry:
                                         position=(row['y_0'], row['x_0']),
                                         mode='trim')[0]] = True
 
-            fit_model = self.fitter(group_psf, x[usepixel], y[usepixel],
-                                    image[usepixel])
+            if uncert is not None:
+                # Uncertainties are defined such that weights should be their
+                # inverse; e.g., for Gaussians this should be 1/sigma.
+                fit_model = self.fitter(group_psf, x[usepixel], y[usepixel],
+                                        image[usepixel], weights=1 / uncert[usepixel])
+            else:
+                fit_model = self.fitter(group_psf, x[usepixel], y[usepixel],
+                                        image[usepixel])
             param_table = self._model_params2table(fit_model,
                                                    len(star_groups.groups[n]))
             result_tab = vstack([result_tab, param_table])
@@ -371,7 +429,7 @@ class BasicPSFPhotometry:
         if 'param_cov' in self.fitter.fit_info.keys():
             result_tab = hstack([result_tab, unc_tab])
 
-        return result_tab, image
+        return result_tab, NDData(image, uncertainty=StdDevUncertainty(uncert))
 
     def _define_fit_param_names(self):
         """
@@ -428,7 +486,7 @@ class BasicPSFPhotometry:
                 n_fit_params = len(unc_tab.colnames)
                 for i in range(star_group_size):
                     unc_tab[i] = np.sqrt(np.diag(
-                                          self.fitter.fit_info['param_cov'])
+                                         self.fitter.fit_info['param_cov'])
                                          )[k: k + n_fit_params]
                     k = k + n_fit_params
         return unc_tab
@@ -540,6 +598,11 @@ class IterativelySubtractedPSFPhotometry(BasicPSFPhotometry):
         stars remain.  Note that in this case it is *possible* that the
         loop will never end if the PSF has structure that causes
         subtraction to create new sources infinitely.
+    noise_calc : callable, optional
+        Function on which uncertainties of data can be calculated. If `None`
+        then `None` is propagated in the uncertainty attribute of the NDData
+        instance and handled accordingly. Must return an array with the same
+        shape as the input array.
 
     Notes
     -----
@@ -558,10 +621,10 @@ class IterativelySubtractedPSFPhotometry(BasicPSFPhotometry):
 
     def __init__(self, group_maker, bkg_estimator, psf_model, fitshape,
                  finder, fitter=LevMarLSQFitter(), niters=3,
-                 aperture_radius=None):
+                 aperture_radius=None, noise_calc=None):
 
         super().__init__(group_maker, bkg_estimator, psf_model, fitshape,
-                         finder, fitter, aperture_radius)
+                         finder, fitter, aperture_radius, noise_calc)
         self.niters = niters
 
     @property
@@ -638,6 +701,20 @@ class IterativelySubtractedPSFPhotometry(BasicPSFPhotometry):
             covariance matrix.
         """
 
+        # Parse image for the different kinds of accepted inputs; if
+        # not ndarray, convert accordingly.
+        if isinstance(image, fits.HDUList):
+            for i in range(len(image)):
+                if image[i].data is not None:
+                    warnings.warn("Input data is a HDUList object, photometry is "
+                                  "run only for the {0} HDU."
+                                  .format(i), AstropyUserWarning)
+                    image = image[i]
+                    break
+        if isinstance(image, (fits.PrimaryHDU, fits.ImageHDU)):
+            image = image.data
+        image = self._make_image(image)
+
         if init_guesses is not None:
             table = super().do_photometry(image, init_guesses)
             table['iter_detected'] = np.ones(table['x_fit'].shape,
@@ -650,7 +727,8 @@ class IterativelySubtractedPSFPhotometry(BasicPSFPhotometry):
             output_table = vstack([table, output_table])
         else:
             if self.bkg_estimator is not None:
-                self._residual_image = image - self.bkg_estimator(image)
+                self._residual_image = NDData(image.data - self.bkg_estimator(image.data),
+                                              uncertainty=StdDevUncertainty(image.uncertainty))
 
             if self.aperture_radius is None:
                 if hasattr(self.psf_model, 'fwhm'):
@@ -693,7 +771,7 @@ class IterativelySubtractedPSFPhotometry(BasicPSFPhotometry):
             output_table.add_column(Column(name=init_parname))
             output_table.add_column(Column(name=fit_parname))
 
-        sources = self.finder(self._residual_image)
+        sources = self.finder(self._residual_image.data)
 
         n = n_start
         while(len(sources) > 0 and
@@ -724,15 +802,15 @@ class IterativelySubtractedPSFPhotometry(BasicPSFPhotometry):
             star_groups = star_groups.group_by('group_id')
             table = hstack([star_groups, table])
 
-            table['iter_detected'] = n*np.ones(table['x_fit'].shape,
-                                               dtype=np.int32)
+            table['iter_detected'] = n * np.ones(table['x_fit'].shape,
+                                                 dtype=np.int32)
 
             output_table = vstack([output_table, table])
 
             # do not warn if no sources are found beyond the first iteration
             with warnings.catch_warnings():
                 warnings.simplefilter('ignore', AstropyUserWarning)
-                sources = self.finder(self._residual_image)
+                sources = self.finder(self._residual_image.data)
 
             n += 1